refactor(project): reorganize folder structure for better maintainability
