+++ next.config.ts
@@ -11,7 +11,7 @@ const nextConfig: NextConfig = {
     return [
       {
         source: '/api/:path*',
-        destination: 'http://localhost:3001/api/:path*', // URL do seu json-server
+        destination: 'http://localhost:3001/api/:path*',
       },
     ];
   },
