+++ src/app/transaction/components/DeleteTransactionModal/DeleteTransactionModal.constants.ts
