+++ src/app/transaction/hooks/useDeleteTransaction.ts
@@ -2,11 +2,12 @@ import { useState, useRef, useCallback, useMemo, useEffect } from 'react';
 import type { Transaction } from '@/types/transaction';
 import { canDeleteTransaction } from '@/utils/validators';
 import { getErrorMessage } from '@/utils/errors';
-import { DELETE_TRANSACTION_MESSAGES } from '@/components/DeleteTransactionModal/DeleteTransactionModal.constants';
+import { DELETE_TRANSACTION_MESSAGES } from '@app/transaction/components/DeleteTransactionModal/DeleteTransactionModal.constants';
 
 interface UseDeleteTransactionOptions {
   onSuccess?: () => void;
   successMessage?: string;
+  successCloseDelay?: number;
 }
 
 interface UseDeleteTransactionReturn {
@@ -58,7 +59,6 @@ export function useDeleteTransaction(
     if (!transaction || !canDelete || isDeleting) return;
 
     setIsDeleting(true);
-    // Limpa mensagens anteriores ao iniciar
     clearMessages();
 
     try {
@@ -68,7 +68,6 @@ export function useDeleteTransaction(
 
       setSuccess(successMessage);
 
-      // A responsabilidade de fechar o modal/limpar a mensagem agora é do componente
       onSuccess?.();
     } catch (err) {
       if (!isMountedRef.current) return;
