+++ src/services/api.ts
@@ -1,6 +1,6 @@
 export const API_CONFIG = {
   BASE_URL: process.env.NEXT_PUBLIC_API_URL || 'http://localhost:3001/api',
-  TIMEOUT: 10000, // 10 segundos
+  TIMEOUT: 10000,
   ENDPOINTS: {
     TRANSACTIONS: '/transactions',
     TRANSACTION_BY_ID: (id: string) => `/transactions/${id}`,
