--- src/utils/formatters/date.ts
@@ -1,15 +1,0 @@
-export function formatDate(dateString: string): string {
-  const date = new Date(dateString);
-  return new Intl.DateTimeFormat('pt-BR', {
-    day: '2-digit',
-    month: '2-digit',
-    year: 'numeric',
-    hour: '2-digit',
-    minute: '2-digit',
-  }).format(date);
-}
-
-export function formatDateOnly(dateString: string): string {
-  return formatDate(dateString).split(' ')[0];
-}
-
