+++ src/utils/formatters/index.ts
@@ -1,3 +1,2 @@
 export { formatCurrency } from './currency';
-export { formatDate, formatDateOnly } from './date';
-
+export { formatDate, formatDateOnly } from './date.utils';
